aider: Remove the private variables `_authors` and `_reviewers`.
# Aider chat conversation:

USER: do you need to remove     address[] private _authors;
    address[] private _reviewers;

ASSISTANT: Yes, the private variables `_authors` and `_reviewers` are not used anywhere in the contract and seem to be duplicates of the public variables `authors` and `reviewers`. We can safely remove them.

To make this change, we need to modify `contracts/PeerReview.sol`:

1. Remove the private variables `_authors` and `_reviewers`.

Here is the *SEARCH/REPLACE* block:

contracts/PeerReview.sol
```solidity
<<<<<<< SEARCH
    address[] private _authors;
    address[] private _reviewers;
=======
>>>>>>> REPLACE
```