aider: The constructor now accepts an array of Reviewer structs instead of an array of addresses.
# Aider chat conversation:

USER: TypeError: Type address[] memory is not implicitly convertible to expected type struct ReviewProcess.Reviewer[] storage ref.
 --> contracts/peerreview.sol:8:21:
  |
8 |         reviewers = _reviewers;
  |                     ^^^^^^^^^^

ASSISTANT: The error is due to the fact that the `_reviewers` parameter in the constructor is of type `address[] memory`, but the `reviewers` state variable is of type `Reviewer[]`.

The `Reviewer` struct contains an `address` and a `string[]` for keywords. So, you cannot directly assign an `address[]` to a `Reviewer[]`.

To fix this, you need to change the constructor to accept an array of `Reviewer` structs instead of an array of addresses.

Here is the *SEARCH/REPLACE* block:

contracts/PeerReview.sol
```solidity
<<<<<<< SEARCH
    constructor(address[] memory _authors, address[] memory _reviewers) {
=======
    constructor(address[] memory _authors, Reviewer[] memory _reviewers) {
>>>>>>> REPLACE
```

Please note that this change will also require modifications to the way the constructor is called, as it now expects an array of `Reviewer` structs, not just addresses.