aider: Fixed copying of Reviewer array from memory to storage in constructor.
# Aider chat conversation:

USER: UnimplementedFeatureError: Copying of type struct ReviewProcess.Reviewer memory[] memory to storage not yet supported.

ASSISTANT: The error you're seeing is because Solidity does not yet support copying an array of structs from memory to storage directly.

To fix this, you need to manually copy each element of the `_reviewers` array to the `reviewers` storage array in the constructor.

Here is the *SEARCH/REPLACE* block:

contracts/PeerReview.sol
```solidity
<<<<<<< SEARCH
    constructor(address[] memory _authors, Reviewer[] memory _reviewers) {
        authors = _authors;
        reviewers = _reviewers;
    }
=======
    constructor(address[] memory _authors, Reviewer[] memory _reviewers) {
        authors = _authors;
        for (uint i = 0; i < _reviewers.length; i++) {
            reviewers.push(_reviewers[i]);
        }
    }
>>>>>>> REPLACE
```

This change will create a new `Reviewer` in the `reviewers` array for each `_reviewer` in the input array, effectively copying the array from memory to storage.