aider: The constructor now accepts an array of addresses for reviewers and a new function allows reviewers to add their keywords.
# Aider chat conversation:

USER: the constractor should take array of addressers for reviewers not array of  Reviewer structs. the reviewers should add the keywords themselves later

ASSISTANT: To implement the requested changes, we need to modify the constructor in the `contracts/PeerReview.sol` file to accept an array of addresses for reviewers instead of an array of `Reviewer` structs. Additionally, we'll introduce a new function that allows reviewers to add their keywords.

Here are the steps and the corresponding *SEARCH/REPLACE* blocks:

1. **Modify the constructor to accept an array of addresses for reviewers.**
2. **Remove the loop that initializes reviewers with keywords in the constructor.**
3. **Add a function that allows reviewers to add their keywords.**

### Step 1 & 2: Modify the Constructor

contracts/PeerReview.sol
```solidity
<<<<<<< SEARCH
    constructor(address[] memory _authors, Reviewer[] memory _reviewers) {
        authors = _authors;
        for (uint i = 0; i < _reviewers.length; i++) {
            reviewers.push(_reviewers[i]);
        }
    }
=======
    constructor(address[] memory _authors, address[] memory _reviewerAddresses) {
        authors = _authors;
        for (uint i = 0; i < _reviewerAddresses.length; i++) {
            reviewers.push(Reviewer(_reviewerAddresses[i], new string[](0)));
        }
    }
>>>>>>> REPLACE
```

### Step 3: Add Function for Reviewers to Add Keywords

contracts/PeerReview.sol
```solidity
<<<<<<< SEARCH
=======
    // Function for reviewers to add their keywords
    function addKeywords(string[] memory _keywords) public {
        for (uint i = 0; i < reviewers.length; i++) {
            if (reviewers[i].addr == msg.sender) {
                reviewers[i].keywords = _keywords;
                return;
            }
        }
        // If the sender is not in the list of reviewers, add them with the provided keywords
        reviewers.push(Reviewer(msg.sender, _keywords));
    }
>>>>>>> REPLACE
```

These changes will adjust the constructor to accept an array of addresses for reviewers and allow reviewers to add their keywords through a new function. This approach provides flexibility for reviewers to update their expertise areas.